feat: Refactored staff section component for About Us page
